python program builds a physical surrogate
This python program builds a physical surrogate model using physical data to replicate a digital twin which is modelled on a Convolutional Neural Network (CNN). It also monitors and continuously updates the model against the surrogate.
The necessary libraries for this algorithm are numpy, tensorflow, and the tensorflow.keras layers library.